Updates the addon version to 11.1.7 and introduces several important fixes and features to improve stability and extend functionality. (#121)
The addon version has been bumped to 11.1.7
Fixed a crash when a spell's tooltip had no owner (could be caused by other addons modifying the game).
Fixed a crash when a quest had no translated title.
Fixed a rare crash that occurred when opening the spellbook.
Implemented a new mechanism to register and translate strings for the context menu.